requirement.txt
fastapi==0.110.1
uvicorn==0.29.0
llama-cpp-python==0.2.57
langdetect==1.0.9
sentence-transformers==2.2.2
chromadb==0.4.24
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
aiofiles==23.2.1
python-multipart==0.0.9
jinja2==3.1.3